Partial refectoring of jobindsats. Added TODOs for additional changes needed and for questions
